chore: no reviews case added, relations separated from schema file to save headache when db introspecting
